Moving unit tests into the unit test folder.
